requirements.txt
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
cycler==0.10.0
idna==2.9
kiwisolver==1.2.0
matplotlib==3.2.1
numpy==1.18.3
Pillow==7.1.1
pynput==1.6.8
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.23.0
six==1.14.0
soupsieve==2.0
urllib3==1.25.9
wxPython==4.0.7.post2
